requirements.txt
django==4.1.2
djangorestframework==3.11
psycopg2
requests
django-jquery
